closes #239 fixing early404
Previously early404 would match the threads with less than "early404Replies" BEFORE skipping and then deleting.
Which would be incorrect because then early404 would only be deleting a fraction of the posts with less than
early404Replies, rather than any thread with less than early404Replies past a certain fraction of the threadLimit

Bit of a stupid mistake, but this should fix it, if my interpretation of how early404 is supposed to work is correct.