Add modlog on unban
Since unbans can be performed on a global ban and there is no certain way to determine the board it was on (e.g. the staff didn't enable Show Post in Ban), a global board option was handled for modlog actions. This could be useful if other global actions are logged in the future, such as modifying global settings, users, filters, etc.

Unfortunately this does mean a global ban will show on the modlog of the board the post was made on, but the global unban will not, if the post wasn't included in the ban. This is not a big concern since by default unbans are not shown on a public modlog, so only the board staff are impacted

This could be avoided by always logging the board of a ban and adding a bool 'global' to the database, although this would require a db migration and some code rewriting.